test: console_recv_short
