+++ examples/tests/console/console_recv_short/README.md
